requirements.txt
altair==4.2.2
appnope==0.1.3
asttokens==2.0.5
attrs==22.2.0
autopep8==1.5.7
backcall==0.2.0
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dateparser==1.1.8
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
fonttools==4.39.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.1.0
ipykernel==6.13.0
ipython==8.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter-client==7.2.2
jupyter-core==4.10.0
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.3
mdurl==0.1.2
mypy==0.971
nest-asyncio==1.5.5
nltk==3.6.1
numpy==1.24.2
packaging==21.3
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.29
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycodestyle==2.7.0
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.8
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
regex==2023.3.23
requests==2.28.2
rich==13.3.3
seaborn==0.12.2
semver==2.13.0
six==1.16.0
smmap==5.0.0
stack-data==0.2.0
streamlit==1.20.0
streamlit-option-menu==0.3.2
toml==0.10.2
toolz==0.12.0
tornado==6.1
tqdm==4.65.0
traitlets==5.1.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
wcwidth==0.2.5
wordcloud==1.8.2.2
zipp==3.15.0
